# spec/api.yaml
openapi: 3.0.0
info:
  title: Movie Poll API
  description: Movie Poll API Service
  version: 0.0.1
  contact:
    name: API Support
    url: https://www.github.com/AdrianPrawda/MoviePoll
servers:
  - url: http://localhost:35555
    description: Local development server
...
paths:
...
  /api/heartbeat:
    post:
      operationId: heartbeat
      tags: [heartbeat]
      summary: Returns a heartbeat
      description: Should always return 200 if the service is operational
      responses:
        '200':
          description: Heartbeat succeeded
        '408':
          description: Heartbeat failed
        default:
          description: Unexpected error